fix: reintroduce explicit type on RHS

+++ test_cases/regression_tests/getbit_fold_oob_bytes.py
@@ -1,10 +1,10 @@
-from algopy import Contract, op
+from algopy import Contract, UInt64, op
 
 
 class GetBitFoldOOBBytes(Contract):
     def approval_program(self) -> bool:
         # out of bounds bit index access (8 bits, indices 0...7)
-        assert op.getbit(b"\xff", 8) == 0
+        assert op.getbit(b"\xff", 8) == UInt64(0)
         return True
 
     def clear_state_program(self) -> bool:
